Abstracted/generalized EventInfo.
